Refactor #185 move VOs
