+++ src/AppBundle/Model/Ip/Tracer/FreeGeoIpTracingService.php
@@ -11,8 +11,7 @@
  */
 
 namespace AppBundle\Model\Ip\Tracer;
-
-use AppBundle\Model\Ip\IpLocation;
+use AppBundle\Model\Ip\Value\IpLocation;
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\ClientException;
 use Psr\Http\Message\StreamInterface;
