+++ src/AppBundle/Tests/Model/Ip/Tracer/FreeGeoIpTracingServiceTest.php
@@ -12,7 +12,7 @@
 
 namespace AppBundle\Tests\Model\Ip\Tracer;
 
-use AppBundle\Model\Ip\IpLocation;
+use AppBundle\Model\Ip\Value\IpLocation;
 use AppBundle\Model\Ip\Tracer\FreeGeoIpTracingService;
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\ClientException;
